Kootenay Lake Exploitation Study Data

Data from the Kootenay Lake Large Trout Exploitation Study.

Details

For a list of the available data sets type data(package = "klexdatr").

For information on a particular dataset type ?dataset.

All spatial coordinates are for UTM Zone 11U, i.e., EPSG: 26911, and all times are in Pacific Standard, i.e. Etc/GMT+8.

Examples

library(dplyr) # so tbl data frames print nice

print(klexdatr::section)

#> # A tibble: 33 x 5
#>    Section Habitat Bounded EastingSection NorthingSection
#>  *  <fctr>  <fctr>   <lgl>          <dbl>           <dbl>
#>  1     S03  Lentic    TRUE       469091.4         5603774
#>  2     S02   Lotic    TRUE       480563.8         5595333
#>  3     S04  Lentic    TRUE       502869.0         5585727
#>  4     S01   Lotic    TRUE       494545.9         5580270
#>  5     S05   Lotic    TRUE       503461.2         5566212
#>  6     S06   Lotic    TRUE       503599.0         5561065
#>  7     S07  Lentic    TRUE       504379.6         5556931
#>  8     S08  Lentic    TRUE       505733.3         5550877
#>  9     S09  Lentic    TRUE       507685.1         5545409
#> 10     S10  Lentic    TRUE       507829.3         5542442
#> # ... with 23 more rows

print(klexdatr::station)

#> # A tibble: 28 x 4
#>                        Station Section EastingStation NorthingStation
#>                         <fctr>  <fctr>          <dbl>           <dbl>
#>  1         Lardeau R. at Mobbs     S02       480847.0         5595246
#>  2          Gerrard Scafolding     S02       480626.0         5595336
#>  3              Gerrard Bridge     S02       480380.0         5595360
#>  4               Lardeau Delta     S07       504927.0         5556704
#>  5                   Fry Creek     S09       507092.4         5545461
#>  6 Kaslo - Shutty Bench - West     S11       507853.0         5538758
#>  7 Kaslo - Shutty Bench - East     S11       507390.0         5538898
#>  8       South of Kaslo - East     S13       508997.0         5524158
#>  9       South of Kaslo - West     S13       508102.0         5524630
#> 10              Woodbury Point     S15       509139.0         5516639
#> # ... with 18 more rows

print(klexdatr::deployment)

#> # A tibble: 197 x 4
#>                                Station Receiver  DateTimeReceiverIn
#>                                 <fctr>   <fctr>              <dttm>
#>  1                  Gerrard Scafolding     2736 2008-04-06 11:00:00
#>  2                      Gerrard Bridge     6247 2008-04-06 11:00:00
#>  3                 Lardeau R. at Mobbs     8247 2008-04-06 11:00:00
#>  4                       Lardeau Delta     6092 2008-03-06 12:00:00
#>  5                       Lardeau Delta     5308 2008-09-16 11:00:00
#>  6         Kaslo - Shutty Bench - West     4348 2008-03-06 12:00:00
#>  7         Kaslo - Shutty Bench - East     5306 2008-03-06 12:00:00
#>  8                      Woodbury Point     8248 2008-05-08 11:00:00
#>  9   Coffee Creek - Power Lines - East     6093 2008-02-26 12:00:00
#> 10 Coffee Creek - Power Lines - Middle     5308 2008-02-26 12:00:00
#> # ... with 187 more rows, and 1 more variables: DateTimeReceiverOut <dttm>

print(klexdatr::capture)

#> # A tibble: 245 x 9
#>    Capture     DateTimeCapture SectionCapture    Species Length Weight Reward1
#>     <fctr>              <dttm>         <fctr>     <fctr>  <int>  <dbl>   <int>
#>  1    F075 2008-05-24 10:40:00            S25 Bull Trout    545   1.75       0
#>  2    F076 2008-05-24 11:10:00            S25 Bull Trout    654   3.40       0
#>  3    F084 2008-06-05 08:01:00            S16 Bull Trout    553     NA       0
#>  4    F100 2009-05-02 13:31:00            S21 Bull Trout    643   2.75     100
#>  5    F104 2009-05-03 11:17:00            S22 Bull Trout    567   2.00     100
#>  6    F129 2009-05-07 09:11:00            S21 Bull Trout    703   3.90     100
#>  7    F133 2009-05-07 13:58:00            S21 Bull Trout    712   3.65     100
#>  8    F137 2009-05-08 11:16:00            S21 Bull Trout    572   1.90     100
#>  9    F141 2009-05-20 15:13:00            S16 Bull Trout    564   2.20     100
#> 10    F143 2009-05-21 08:00:00            S16 Bull Trout    595   2.40     100
#> # ... with 235 more rows, and 2 more variables: Reward2 <int>,
#> #   DateTimeTagExpire <dttm>

print(klexdatr::recapture)

#> # A tibble: 54 x 8
#>      DateTimeRecapture Capture SectionRecapture TBarTag1 TBarTag2 TagsRemoved
#>                 <dttm>  <fctr>           <fctr>    <lgl>    <lgl>       <lgl>
#>  1 2009-04-05 12:00:00    F006              S25     TRUE     TRUE        TRUE
#>  2 2009-05-23 12:00:00    F131              S16     TRUE     TRUE        TRUE
#>  3 2009-09-10 12:00:00    F097              S14     TRUE     TRUE        TRUE
#>  4 2009-10-06 12:00:00    F169              S08     TRUE     TRUE        TRUE
#>  5 2009-10-10 12:00:00    F076              S30     TRUE     TRUE        TRUE
#>  6 2009-10-12 12:00:00    F099              S10     TRUE     TRUE        TRUE
#>  7 2009-10-14 12:00:00    F112              S16     TRUE     TRUE        TRUE
#>  8 2009-10-16 12:00:00    F111              S27     TRUE     TRUE        TRUE
#>  9 2010-01-23 12:00:00    F197              S27     TRUE     TRUE        TRUE
#> 10 2010-01-28 12:00:00    F075              S23     TRUE     TRUE        TRUE
#> # ... with 44 more rows, and 2 more variables: Released <lgl>, Public <lgl>

print(klexdatr::detection)

#> # tibble [313,522 x 4]
#>      DateTimeDetection Capture Receiver Detections
#>                 <dttm>  <fctr>   <fctr>      <int>
#>  1 2012-12-30 18:00:00    F379   220102          5
#>  2 2012-12-31 17:00:00    F379   220009         14
#>  3 2012-12-31 18:00:00    F379   220009         11
#>  4 2012-12-31 19:00:00    F379   220009          5
#>  5 2012-12-31 22:00:00    F379   220009          4
#>  6 2012-12-31 23:00:00    F379   220009          3
#>  7 2013-01-01 05:00:00    F379   220009          3
#>  8 2013-01-02 02:00:00    F379   220009          7
#>  9 2013-01-02 15:00:00    F379   220009          5
#> 10 2013-01-03 09:00:00    F379   220125          3
#> # ... with 313,512 more rows